+++ src/vim-loader/vim.ts
@@ -1,140 +1,208 @@
-import { BFast } from './bfast'
-import { G3d } from './g3d'
+import * as THREE from 'three'
+import { Document } from './document'
+import { Scene } from './scene'
 
-export type EntityTable = Map<string, ArrayLike<number>>
-
+/**
+ * Container for the built three meshes and the vim data from which it was built.
+ * Maps between BIM data and Three objects
+ * Provides an interface to access BIM data.
+ */
 export class Vim {
-  static tableElement = 'Vim.Element'
-  static tableElementLegacy = 'Rvt.Element'
-  static tableNode = 'Vim.Node'
+  vim: Document
+  scene: Scene
+  elementIndexToInstanceIndices: Map<number, number[]>
+  elementIdToElementIndex: Map<number, number>
 
-  header: string
-  assets: BFast
-  g3d: G3d
-  entities: Map<string, EntityTable>
-  strings: string[]
-
-  constructor (
-    header: string,
-    assets: BFast,
-    g3d: G3d,
-    entities: Map<string, EntityTable>,
-    strings: string[]
-  ) {
-    this.header = header
-    this.assets = assets
-    this.g3d = g3d
-    this.entities = entities
-    this.strings = strings
+  constructor (vim: Document, scene: Scene) {
+    this.vim = vim
+    this.scene = scene
+    this.elementIndexToInstanceIndices = this.mapElementIndexToInstanceIndices()
+    this.elementIdToElementIndex = this.mapElementIdToIndex()
   }
 
-  getEntity (type: string, index: number): any {
-    const r = new Map<string, string | number>()
-    if (index < 0) return r
-    const table = this.entities?.get(type)
-    if (!table) return r
-    for (const k of table.keys()) {
-      const parts = k.split(':')
-      const values = table.get(k)
-      if (!values) continue
+  private mapElementIndexToInstanceIndices (): Map<number, number[]> {
+    const map = new Map<number, number[]>()
+    const nodeElements = this.getElementIndices(this.getNodeTable())
+    const nodeCount = nodeElements.length
 
-      const value =
-        parts[0] === 'string' ? this.strings[values[index]] : values[index]
+    for (let node = 0; node < nodeCount; node++) {
+      const element = nodeElements[node]
+      if (element === undefined) continue
 
-      const name = parts[parts.length - 1]
-      r.set(name, value)
+      const nodes = map.get(element)
+      if (nodes) {
+        nodes.push(node)
+      } else {
+        map.set(element, [node])
+      }
     }
-    return r
+    return map
   }
 
-  static parseFromArrayBuffer (data: ArrayBuffer) {
-    const bfast = BFast.fromArrayBuffer(data)
-    return Vim.parseFromBFast(bfast)
-  }
+  private mapElementIdToIndex (): Map<number, number> {
+    const map = new Map<number, number>()
+    const elementIds = this.getIntColumn(this.getElementTable(), 'Id')
 
-  // Given a BFAST container (header/names/buffers) constructs a VIM data structure
-  static parseFromBFast (bfast: BFast): Vim {
-    if (bfast.buffers.length < 5) {
-      throw new Error('VIM requires at least five BFast buffers')
-    }
+    let negativeReported = false
+    let duplicateReported = false
+    for (let element = 0; element < elementIds.length; element++) {
+      const id = elementIds[element]
 
-    const lookup = new Map<string, any>()
-    for (let i = 0; i < bfast.buffers.length; ++i) {
-      lookup.set(bfast.names[i], bfast.buffers[i])
-    }
+      if (id < 0) {
+        if (!negativeReported) {
+          console.error('Ignoring negative element ids. Check source data.')
+          negativeReported = true
+        }
 
-    const assetData = lookup.get('assets')
-    const g3dData = lookup.get('geometry')
-    const headerData = lookup.get('header')
-    const entityData = lookup.get('entities')
-    const stringData = lookup.get('strings')
-
-    const header = new TextDecoder('utf-8').decode(headerData)
-    const g3d = G3d.fromBfast(BFast.fromArray(g3dData))
-    const assets = BFast.fromArray(assetData)
-    const entities = Vim.parseEntityTables(BFast.fromArray(entityData))
-    const strings = new TextDecoder('utf-8').decode(stringData).split('\0')
-
-    g3d.validate()
-
-    return new Vim(header, assets, g3d, entities, strings)
-  }
-
-  static parseEntityTables (bfast: BFast): Map<string, EntityTable> {
-    const result = new Map<string, any>()
-    for (let i = 0; i < bfast.buffers.length; ++i) {
-      const current = bfast.names[i]
-      const tableName = current.substring(current.indexOf(':') + 1)
-      const buffer = bfast.buffers[i]
-      const next = Vim.parseEntityTable(BFast.fromArray(buffer))
-      result.set(tableName, next)
-    }
-    return result
-  }
-
-  static parseEntityTable (bfast: BFast): EntityTable {
-    const result = new Map<string, any>()
-    for (let i = 0; i < bfast.buffers.length; ++i) {
-      const columnName = bfast.names[i]
-      // eslint-disable-next-line no-unused-vars
-      const [columnType, ..._] = columnName.split(':')
-      const buffer = bfast.buffers[i]
-
-      let length: number
-      let ctor:
-        | Int8ArrayConstructor
-        | Float32ArrayConstructor
-        | Float64ArrayConstructor
-        | Int32ArrayConstructor
-      switch (columnType) {
-        case 'byte':
-          length = buffer.byteLength
-          ctor = Int8Array
-          break
-        case 'float':
-          length = buffer.byteLength / 4
-          ctor = Float32Array
-          break
-        case 'double':
-        case 'numeric': // legacy (vim0.9)
-          length = buffer.byteLength / 8
-          ctor = Float64Array
-          break
-        case 'string': // i.e. indices into the string table
-        case 'index':
-        case 'int':
-        case 'properties': // legacy (vim0.9)
-          length = buffer.byteLength / 4
-          ctor = Int32Array
-          break
-        default:
-          throw new Error('Unrecognized column type ' + columnType)
+        continue
+      }
+      if (map.has(id)) {
+        if (!duplicateReported) {
+          console.error('Ignoring duplicate element ids. Check source data.')
+          duplicateReported = true
+          continue
+        }
       }
 
-      // eslint-disable-next-line new-cap
-      const columnData = new ctor(buffer.buffer, buffer.byteOffset, length)
-      result.set(columnName, columnData)
+      map.set(id, element)
     }
+    return map
+  }
+
+  getElementIndexFromElementId = (elementId: number): number | undefined =>
+    this.elementIdToElementIndex.get(elementId)
+
+  getStringColumn = (table: any, colNameNoPrefix: string): number[] =>
+    table?.get('string:' + colNameNoPrefix)
+
+  getIndexColumn = (table: any, tableName: string, fieldName: string) =>
+    table?.get(`index:${tableName}:${fieldName}`)
+
+  getDataColumn = (table: any, typePrefix: any, colNameNoPrefix: any) =>
+    table?.get(typePrefix + colNameNoPrefix) ??
+    table?.get('numeric:' + colNameNoPrefix) // Backwards compatible call with vim0.9
+
+  getIntColumn = (table: any, colNameNoPrefix: string) =>
+    this.getDataColumn(table, 'int:', colNameNoPrefix)
+
+  getByteColumn = (table: any, colNameNoPrefix: string) =>
+    this.getDataColumn(table, 'byte:', colNameNoPrefix)
+
+  getFloatColumn = (table: any, colNameNoPrefix: string) =>
+    this.getDataColumn(table, 'float:', colNameNoPrefix)
+
+  getDoubleColumn = (table: any, colNameNoPrefix: string) =>
+    this.getDataColumn(table, 'double:', colNameNoPrefix)
+
+  getElementIndices = (table: any): number[] =>
+    this.getIndexColumn(table, Document.tableElement, 'Element') ??
+    table?.get(Document.tableElementLegacy) // Backwards compatible call with vim0.9
+
+  getElementTable = () =>
+    this.vim.entities?.get(Document.tableElement) ??
+    this.vim.entities?.get(Document.tableElementLegacy)
+
+  getNodeTable = () => this.vim.entities.get(Document.tableNode)
+
+  /**
+   * Get Node/Instance Indices for given element index
+   * @param elementIndex element index for which to get node indices
+   * @returns array of node indices or undefined if no corresponding nodes
+   */
+  getNodeIndicesFromElementIndex (elementIndex: number): number[] | undefined {
+    return this.elementIndexToInstanceIndices.get(elementIndex)
+  }
+
+  /**
+   * Get Node/Instance Indices for given element indices
+   * @param elementIndex element indices for which to get node indices
+   * @returns array of node indices, can be empty if no matching nodes
+   */
+  getNodeIndicesFromElementIndices (elementIndices: number[]): number[] {
+    return elementIndices
+      .flatMap((e) => this.getNodeIndicesFromElementIndex(e))
+      .filter((n): n is number => n !== undefined)
+  }
+
+  /**
+   * Get Instance Indices for given element ids
+   * Throws error if argument is undefined
+   * @param elementIds element ids for which to get node indices
+   * @returns array of node indices, can be empty if no matching nodes
+   */
+  getInstanceIndicesFromElementIds (elementIds: number[]): number[] {
+    if (!elementIds) throw new Error('undefined argument')
+
+    // element ids -> element indices
+    const elementIndices = elementIds
+      .map((id) => this.getElementIndexFromElementId(id))
+      .filter((i): i is number => i !== undefined)
+
+    // element indices -> nodes indices
+    return this.getNodeIndicesFromElementIndices(elementIndices)
+  }
+
+  getMeshesFromElementIndex (
+    elementIndex: number
+  ): [THREE.Mesh, number][] | null {
+    const nodeIndices = this.getNodeIndicesFromElementIndex(elementIndex)
+    if (!nodeIndices || !nodeIndices.length) return null
+
+    const result: [THREE.Mesh, number][] = []
+    nodeIndices.forEach((i) => {
+      const mesh = this.getMeshFromNodeIndex(i)
+      if (mesh) result.push(mesh)
+    })
     return result
+  }
+
+  getMeshFromNodeIndex (nodeIndex: number): [THREE.Mesh, number] | undefined {
+    if (nodeIndex < 0) throw new Error('Invalid negative index')
+    const array = this.scene.instanceToThreeMesh.get(nodeIndex)
+    return array ? array[0] : undefined
+  }
+
+  getNodeIndexFromMesh (mesh: THREE.Mesh, instance: number): number {
+    if (!mesh || instance < 0) return -1
+    const nodes = this.scene.threeMeshIdToInstance.get(mesh.id)
+    if (!nodes) return -1
+    return nodes[instance]
+  }
+
+  getElementIndexFromMesh (mesh: THREE.Mesh, instance: number): number {
+    if (!mesh || instance < 0) return -1
+    const nodeIndex = this.getNodeIndexFromMesh(mesh, instance)
+    return this.getElementIndexFromNodeIndex(nodeIndex)
+  }
+
+  getElementIndexFromNodeIndex (nodeIndex: number): number {
+    if (nodeIndex < 0) return -1
+    const node = this.getNodeTable()
+    if (!node) return -1
+    const elements = this.getElementIndices(node)
+    if (!elements) return -1
+    return elements[nodeIndex]
+  }
+
+  getElementIdFromNodeIndex (nodeIndex: number): number {
+    if (nodeIndex < 0) return -1
+    const elementIndex = this.getElementIndexFromNodeIndex(nodeIndex)
+    if (elementIndex < 0) return -1
+    const ids = this.getIntColumn(this.getElementTable(), 'Id')
+    if (!ids) return -1
+    return ids[elementIndex]
+  }
+
+  // TODO add better ways to access bim
+  getElementName (elementIndex: number): string {
+    if (elementIndex < 0) return ''
+    const names = this.getStringColumn(this.getElementTable(), 'Name')
+    if (!names) return ''
+    const nameIndex = names[elementIndex] as number
+    return this.getStringFromIndex(nameIndex)
+  }
+
+  getStringFromIndex (stringIndex: number): string {
+    return stringIndex < 0 ? '' : this.vim.strings[stringIndex]
   }
 }
