+++ src/vim-webgl-viewer.ts
@@ -15,8 +15,8 @@ export * from './vim-webgl-viewer/hitTester'
 
 export * from './vim-loader/bfast'
 export * from './vim-loader/g3d'
+export * from './vim-loader/document'
+export * from './vim-loader/vimLoader'
 export * from './vim-loader/vim'
-export * from './vim-loader/VIMLoader'
-export * from './vim-loader/bimModel'
-export * from './vim-loader/model'
+export * from './vim-loader/scene'
 export * from './vim-loader/geometry'
