+++ src/vim-webgl-viewer/viewerRenderer.ts
@@ -1,5 +1,5 @@
 import * as THREE from 'three'
-import { Model } from '../vim-webgl-viewer'
+import { Scene } from '../vim-loader/scene'
 import { ViewerSettings } from './viewerSettings'
 
 export class ViewerRenderer {
@@ -83,7 +83,7 @@ export class ViewerRenderer {
     this.camera.updateProjectionMatrix()
   }
 
-  clearModels () {
+  clearScene () {
     this.meshes.forEach((m) => this.scene.remove(m))
     this.meshes = []
     this.localBoundingBox = undefined
@@ -106,15 +106,15 @@ export class ViewerRenderer {
     if (i > 0) this.meshes.splice(i, 1)
   }
 
-  addModel (model: Model) {
-    model.meshes.forEach((m) => {
+  addScene (scene: Scene) {
+    scene.meshes.forEach((m) => {
       this.scene.add(m)
       this.meshes.push(m)
     })
 
     this.localBoundingBox = this.localBoundingBox
-      ? this.localBoundingBox.union(model.boundingBox)
-      : model.boundingBox.clone()
+      ? this.localBoundingBox.union(scene.boundingBox)
+      : scene.boundingBox.clone()
 
     this.worldBoundingBox = this.worldBoundingBox ?? this.localBoundingBox
   }
