+++ src/vim-webgl-viewer/viewerSettings.ts
@@ -5,10 +5,7 @@
 import * as THREE from 'three'
 import deepmerge from 'deepmerge'
 import { clone, cloneDeep } from 'lodash'
-import {
-  transparencyIsValid,
-  TransparencyMode
-} from '../vim-loader/geometry'
+import { transparencyIsValid, TransparencyMode } from '../vim-loader/geometry'
 
 export type Vector3 = {
   x: number
@@ -29,10 +26,10 @@ export type ColorHSL = {
 }
 
 /**
- * Plane under model related options
+ * Plane under Scene related options
  */
 export type PlaneOptions = {
-  /** Enables/Disables plane under model */
+  /** Enables/Disables plane under scene */
   show: boolean
   /** Local or remote texture url for plane */
   texture: string
@@ -40,7 +37,7 @@ export type PlaneOptions = {
   opacity: number
   /** Color of the plane */
   color: ColorRGB
-  /** Actual size is ModelRadius*size */
+  /** Actual size is SceneRadius*size */
   size: number
 }
 
@@ -61,8 +58,8 @@ export type CameraControlsOptions = {
    * <p>Orbit rotates the camera around a focus point</p>
    */
   orbit: boolean
-  /** Camera speed is scaled according to modelRadius/modelReferenceSize */
-  modelReferenceSize: number
+  /** Camera speed is scaled according to SceneRadius/sceneReferenceSize */
+  vimReferenceSize: number
   /** Camera rotation speed factor */
   rotateSpeed: number
   orbitSpeed: number
@@ -97,7 +94,7 @@ export type SkyLightOptions = {
   intensity: number
 }
 
-/** Viewer related options independant from models */
+/** Viewer related options independant from vims */
 export type ViewerOptions = {
   /**
    * Webgl canvas related options
@@ -109,7 +106,7 @@ export type ViewerOptions = {
   camera: Partial<CameraOptions>
   // background: Partial<BackgroundOptions>
   /**
-   * Plane under model related options
+   * Plane under scene related options
    */
   plane: Partial<PlaneOptions>
   /**
@@ -123,23 +120,23 @@ export type ViewerOptions = {
 }
 
 /**
- * Config object for loading a model
+ * Config object for loading a vim
  */
-export type ModelOptions = {
+export type VimOptions = {
   /**
-   * Local or remote url of model to load
+   * Local or remote url of vim to load
    */
   url: string
   /**
-   * Position offset for the model
+   * Position offset for the vim
    */
   position: Vector3
   /**
-   * Rotation for the model
+   * Rotation for the vim
    */
   rotation: Vector3
   /**
-   * Scale factor for the model
+   * Scale factor for the vim
    */
   scale: number
   /**
@@ -151,15 +148,15 @@ export type ModelOptions = {
 }
 
 /**
- * <p>Wrapper around Model Options.</p>
+ * <p>Wrapper around Vim Options.</p>
  * <p>Casts options values into related THREE.js type</p>
  * <p>Provides default values for options</p>
  */
-export class ModelSettings {
-  private options: ModelOptions
+export class VimSettings {
+  private options: VimOptions
 
-  constructor (options?: Partial<ModelOptions>) {
-    const fallback: ModelOptions = {
+  constructor (options?: Partial<VimOptions>) {
+    const fallback: VimOptions = {
       url: 'https://vim.azureedge.net/samples/residence.vim',
       position: { x: 0, y: 0, z: 0 },
       rotation: { x: 0, y: 0, z: 0 },
@@ -174,18 +171,17 @@ export class ModelSettings {
       : 'all'
   }
 
-  getOptions = () => cloneDeep(this.options) as ModelOptions
+  getOptions = () => cloneDeep(this.options) as VimOptions
   getURL = () => this.options.url
 
-  // Model
-  getModelPosition = () => toVec(this.options.position)
-  getModelRotation = () => toQuaternion(this.options.rotation)
-  getModelScale = () => scalarToVec(this.options.scale)
-  getModelMatrix = () =>
+  getPosition = () => toVec(this.options.position)
+  getRotation = () => toQuaternion(this.options.rotation)
+  getScale = () => scalarToVec(this.options.scale)
+  getMatrix = () =>
     new THREE.Matrix4().compose(
-      this.getModelPosition(),
-      this.getModelRotation(),
-      this.getModelScale()
+      this.getPosition(),
+      this.getRotation(),
+      this.getScale()
     )
 
   getElementIdsFilter = () => clone(this.options.elementIds)
@@ -213,7 +209,7 @@ export class ViewerSettings {
         zoom: 1,
         controls: {
           orbit: true,
-          modelReferenceSize: 1,
+          vimReferenceSize: 1,
           rotateSpeed: 1,
           orbitSpeed: 1,
           moveSpeed: 1
@@ -275,8 +271,8 @@ export class ViewerSettings {
   getCameraMoveSpeed = () => this.options.camera.controls.moveSpeed
   getCameraRotateSpeed = () => this.options.camera.controls.rotateSpeed
   getCameraOrbitSpeed = () => this.options.camera.controls.orbitSpeed
-  getCameraReferenceModelSize = () =>
-    this.options.camera.controls.modelReferenceSize
+  getCameraReferenceVimSize = () =>
+    this.options.camera.controls.vimReferenceSize
 }
 
 function toRGBColor (c: ColorRGB): THREE.Color {
